Added files via upload
As the code starts running, the orange light is briefly on switching right over to the red light. The red LED stays on for over 7 seconds and then there is under 3 seconds where both red and orange LEDs turn on. Finally, the green LED turns on.

The goal of this project is to get a basic understanding of the coding environment while keeping the challenge level relatively low.